fix: address react-doctor lint findings
Split options/FunctionList.tsx into FunctionListItem.tsx and
FunctionItemParts.tsx to satisfy no-multi-comp, remove unused exports
(ColorPicker, DividerH, filterFunctions), fix duplicate #FFC107 typo in
colorPalette to #FF9800, move render-time ref mutations into
useLayoutEffect, fix exhaustive-deps, parallelize getActiveTab and
repository.get with Promise.all, and switch the popup keydown listener
to useEffectEvent.
