+++ .node-version
@@ -1,1 +1,1 @@
-24.18.1
+26.4.0
